analytics/requirements.txt
pyspark==2.4.3
matplotlib
pandas
sklearn
statsmodels
cassandra-driver